# cmd/projectmanager/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/repository": {
            "get": {
                "description": "list current repository",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "list current repository",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.ListRepositoryResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "add new repository",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "add new repository",
                "parameters": [
                    {
                        "description": "Repository 등록",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gin.AddRepositoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.AddRepositoryResponse"
                        }
                    }
                }
            }
        },
        "/api/repository/delete/{repositoryId}": {
            "delete": {
                "description": "delete repository",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "delete repository",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repositoryId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Repository 삭제",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gin.DeleteRepositoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.DeleteRepositoryResponse"
                        }
                    }
                }
            }
        },
        "/api/repository/detail/{repositoryId}": {
            "get": {
                "description": "get repository info by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "get repository info by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository ID",
                        "name": "repositoryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.GetRepositoryResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gin.AddRepositoryRequest": {
            "type": "object",
            "required": [
                "account_id",
                "repository_name"
            ],
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "repository_name": {
                    "type": "string"
                }
            }
        },
        "gin.AddRepositoryResponse": {
            "type": "object",
            "properties": {
                "repository_id": {
                    "type": "string"
                }
            }
        },
...
        "gin.DeleteRepositoryRequest": {
            "type": "object"
        },
        "gin.DeleteRepositoryResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "gin.GetRepositoryResponse": {
            "type": "object",
            "properties": {
                "repository": {
                    "$ref": "#/definitions/gin.Repository"
                }
            }
        },
...
        "gin.ListRepositoryResponse": {
            "type": "object",
            "properties": {
                "repositories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gin.Repository"
                    }
                }
            }
        },
        "gin.Repository": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}